requirements.txt
dj-database-url==0.5.0
Django==4.0.5
django-environ==0.9.0
django-rest-framework==0.1.0
drf-spectacular==0.22.1
factory-boy==3.2.1
